Made views into a package.
Files checked in: DataRepo/urls.py DataRepo/views.py DataRepo/views/__init__.py DataRepo/views/loading/submission.py DataRepo/views/loading/validation.py DataRepo/views/models/__init__.py DataRepo/views/models/animal.py DataRepo/views/models/compound.py DataRepo/views/models/msrun.py DataRepo/views/models/peakdata.py DataRepo/views/models/peakgroup.py DataRepo/views/models/peakgroupset.py DataRepo/views/models/protocol.py DataRepo/views/models/sample.py DataRepo/views/models/study.py DataRepo/views/models/tissue.py DataRepo/views/nav.py DataRepo/views/search/advanced/download.py DataRepo/views/search/advanced/view.py DataRepo/views/search/basic.py DataRepo/views/search/results.py DataRepo/views/utils.py
