+++ DataRepo/views/models/animal.py
@@ -1,0 +1,42 @@
+from DataRepo.models import Animal
+from django.views.generic import DetailView, ListView
+from DataRepo.utils import QuerysetToPandasDataFrame as qs2df
+
+
+class AnimalListView(ListView):
+    """Generic class-based view for a list of animals"""
+
+    model = Animal
+    context_object_name = "animal_list"
+    template_name = "DataRepo/animal_list.html"
+    ordering = ["name"]
+
+    def get_context_data(self, **kwargs):
+        # Call the base implementation first to get the context
+        context = super(AnimalListView, self).get_context_data(**kwargs)
+        # add data from the DataFrame to the context
+        anim_list_stats_df = qs2df.get_animal_list_stats_df()
+
+        # convert DataFrame to a list of dictionary
+        data = qs2df.df_to_list_of_dict(anim_list_stats_df)
+        context["df"] = data
+        return context
+
+
+class AnimalDetailView(DetailView):
+    """Generic class-based detail view for an animal"""
+
+    model = Animal
+    template_name = "DataRepo/animal_detail.html"
+
+    def get_context_data(self, **kwargs):
+        # Call the base implementation first to get the context
+        context = super(AnimalDetailView, self).get_context_data(**kwargs)
+
+        pk = self.kwargs.get("pk")
+        per_anim_msrun_df = qs2df().get_per_animal_msrun_df(pk)
+
+        # convert DataFrame to a list of dictionary
+        data = qs2df.df_to_list_of_dict(per_anim_msrun_df)
+        context["df"] = data
+        return context
