+++ DataRepo/views/models/msrun.py
@@ -1,0 +1,19 @@
+from DataRepo.models import MSRun
+from django.views.generic import DetailView, ListView
+
+
+class MSRunListView(ListView):
+    """Generic class-based view for a list of MS runs"""
+
+    model = MSRun
+    context_object_name = "msrun_list"
+    template_name = "DataRepo/msrun_list.html"
+    ordering = ["id"]
+    paginate_by = 20
+
+
+class MSRunDetailView(DetailView):
+    """Generic class-based detail view for a MS run"""
+
+    model = MSRun
+    template_name = "DataRepo/msrun_detail.html"
