+++ DataRepo/views/models/peakdata.py
@@ -1,0 +1,27 @@
+from DataRepo.models import PeakGroup, PeakData
+from django.views.generic import ListView
+from django.shortcuts import get_object_or_404
+
+
+class PeakDataListView(ListView):
+    """
+    Generic class-based view for a list of peak data
+    "model = PeakData" is shorthand for queryset = PeakData.objects.all()
+    use queryset syntax for PeakData list with or without filtering
+    """
+
+    queryset = PeakData.objects.all()
+    context_object_name = "peakdata_list"
+    template_name = "DataRepo/peakdata_list.html"
+    ordering = ["peak_group_id", "id"]
+    paginate_by = 200
+
+    # filter peakgdata_list by peak_group_id
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        # get query string from request
+        peakgroup_pk = self.request.GET.get("peak_group_id", None)
+        if peakgroup_pk is not None:
+            self.peakgroup = get_object_or_404(PeakGroup, id=peakgroup_pk)
+            queryset = PeakData.objects.filter(peak_group_id=peakgroup_pk)
+        return queryset
