+++ DataRepo/views/models/peakgroupset.py
@@ -1,0 +1,19 @@
+from DataRepo.models import PeakGroupSet
+from django.views.generic import DetailView, ListView
+
+
+class PeakGroupSetListView(ListView):
+    """Generic class-based view for a list of PeakGroup sets"""
+
+    model = PeakGroupSet
+    context_object_name = "peakgroupset_list"
+    template_name = "DataRepo/peakgroupset_list.html"
+    ordering = ["id"]
+    paginate_by = 20
+
+
+class PeakGroupSetDetailView(DetailView):
+    """Generic class-based detail view for a PeakGroup set"""
+
+    model = PeakGroupSet
+    template_name = "DataRepo/peakgroupset_detail.html"
