+++ DataRepo/views/models/protocol.py
@@ -1,0 +1,19 @@
+from DataRepo.models import Protocol
+from django.views.generic import DetailView, ListView
+
+
+class ProtocolListView(ListView):
+    """Generic class-based view for a list of protocols"""
+
+    model = Protocol
+    context_object_name = "protocol_list"
+    template_name = "DataRepo/protocol_list.html"
+    ordering = ["name"]
+    paginate_by = 20
+
+
+class ProtocolDetailView(DetailView):
+    """Generic class-based detail view for a protocol"""
+
+    model = Protocol
+    template_name = "DataRepo/protocol_detail.html"
